Initial Commit (demo)
This is the initial commit of the web page that will have a CompTIA Network+ practice test.

Every time the user should encounter different questions randomly chosen from the Pools of questions created.

Each pool (only one for the demo) should challenge the user in a different domain of the CompTIA Net+ exam.

In the demo, the user will have to answer 5 questions (from a pool of 15).
The user passes if he gets 80% or more correct.